Move RageTexture to RageUtil/Graphics/Texture
