+++ src/Etterna/Actor/Base/ModelTypes.cpp
@@ -4,8 +4,8 @@
 #include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "RageUtil/File/RageFile.h"
 #include "RageUtil/Misc/RageMath.h"
-#include "RageUtil/Graphics/RageTexture.h"
-#include "RageUtil/Graphics/RageTextureManager.h"
+#include "RageUtil/Graphics/Display/Texture/RageTexture.h"
+#include "RageUtil/Graphics/Display/Texture/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 
 #include <cstring>
