+++ src/Etterna/Actor/Gameplay/Foreground.cpp
@@ -3,7 +3,7 @@
 #include "Etterna/Models/Misc/BackgroundUtil.h"
 #include "Foreground.h"
 #include "Etterna/Singletons/GameState.h"
-#include "RageUtil/Graphics/RageTextureManager.h"
+#include "RageUtil/Graphics/Display/Texture/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Etterna/Models/Songs/Song.h"
 #include "Etterna/Models/Songs/SongOptions.h"
