+++ src/Etterna/Actor/Menus/GraphDisplay.cpp
@@ -4,7 +4,7 @@
 #include "GraphDisplay.h"
 #include "RageUtil/Graphics/Display/RageDisplay.h"
 #include "RageUtil/Misc/RageMath.h"
-#include "RageUtil/Graphics/RageTextureManager.h"
+#include "RageUtil/Graphics/Display/Texture/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Etterna/Models/Songs/Song.h"
 #include "Etterna/Models/Misc/StageStats.h"
