+++ src/Etterna/Models/Misc/ImageCache.cpp
@@ -10,8 +10,8 @@
 #include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Palettize.h"
 #include "RageUtil/Graphics/Display/Surface/RageSurfaceUtils_Zoom.h"
 #include "RageUtil/Graphics/Display/Surface/RageSurface_Load.h"
-#include "RageUtil/Graphics/RageTexture.h"
-#include "RageUtil/Graphics/RageTextureManager.h"
+#include "RageUtil/Graphics/Display/Texture/RageTexture.h"
+#include "RageUtil/Graphics/Display/Texture/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "CommonMetrics.h"
 #include "Etterna/Models/Songs/SongCacheIndex.h"
