+++ src/RageUtil/Graphics/Display/Texture/RageTexture.cpp
