+++ src/RageUtil/Graphics/Display/Texture/RageTexture.h
