+++ src/RageUtil/Graphics/Display/Texture/RageTextureID.h
