+++ src/RageUtil/Graphics/Display/Texture/RageTextureManager.cpp
