+++ src/RageUtil/Graphics/Display/Texture/RageTextureManager.h
