+++ src/RageUtil/Graphics/Display/Texture/RageTextureRenderTarget.cpp
