+++ src/arch/MovieTexture/MovieTexture_Null.cpp
@@ -1,6 +1,6 @@
 #include "Etterna/Globals/global.h"
 #include "RageUtil/Graphics/Display/RageDisplay.h"
-#include "RageUtil/Graphics/RageTextureManager.h"
+#include "RageUtil/Graphics/Display/Texture/RageTextureManager.h"
 #include "RageUtil/Utils/RageUtil.h"
 #include "Core/Services/Locator.hpp"
 #include "MovieTexture_Null.h"
